requirements.txt
# SCA2LSTM 水文预测系统依赖包
# Python 3.10 + PyTorch 2.9.1 + CUDA 12.6 优化版本
# ==========================================

# 基础科学计算 (Python 3.10 兼容版本)
numpy>=1.24.0              # 1.24+ 完全支持PyTorch 2.9.1
pandas>=1.5.0               # 1.5+ 支持Python 3.10，兼容PyTorch 2.x
scikit-learn>=1.2.0         # 1.2+ 支持Python 3.10和PyTorch 2.x
scipy>=1.9.0                # 1.9+ 支持Python 3.10

# 深度学习框架 (PyTorch 2.9.1 + CUDA 12.6)
# 注意：PyTorch 2.9.1已安装，以下版本确保兼容性
torchvision>=0.14.0         # 与PyTorch 2.9.1配套版本

# 数据可视化 (PyTorch 2.9.1 兼容版本)
matplotlib>=3.6.0           # 3.6+ 支持PyTorch 2.x和Python 3.10
seaborn>=0.12.0              # 0.12+ 支持PyTorch 2.x
plotly>=5.15.0               # 5.15+ 支持PyTorch 2.x

# 进度条和并行处理
tqdm>=4.64.0                 # 4.64+ 支持PyTorch 2.x
joblib>=1.2.0                # 1.2+ 支持Python 3.10和PyTorch 2.x

# 配置文件处理
pyyaml>=6.0.0                # 6.0+ 支持Python 3.10

# 性能优化
numba>=0.57.0                # 0.57+ 支持Python 3.10和PyTorch 2.x

# 时间序列处理
statsmodels>=0.14.0           # 0.14+ 支持Python 3.10和PyTorch 2.x

# 数据预处理
openpyxl>=3.1.0              # 3.1+ 支持Python 3.10